llm_intelligence/pyproject.toml
[project]
name = "llm-intelligence"
version = "0.1.0"
description = "LLM-powered meeting transcript insights microservice"
readme = "README.md"
authors = [
    { name = "Krzysztof Jurkiewicz", email = "huntekah@gmail.com" }
]
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.6.0",
    "google-genai>=1.0.0",
    "ollama>=0.4.0",
    "diskcache>=5.6.0",
    "loguru>=0.7.0",
    "python-dotenv>=1.0.0",
    "tenacity>=9.0.0",
]

[dependency-groups]
dev = [
    "ruff>=0.8.0",
    "bandit>=1.7.0",
    "mypy>=1.13.0",
    "radon>=6.0.0",
    "vulture>=2.0.0",
    "pylint>=3.0.0",
    "pytest>=9.0.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.27.0",
    "pytest-cov>=6.0.0",
    "pytest-mock>=3.12.0",
]

[build-system]
requires = ["uv_build>=0.10.2,<0.11.0"]
build-backend = "uv_build"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
pythonpath = ["src"]
filterwarnings = [
    "ignore::DeprecationWarning",
]

[tool.mypy]
python_version = "3.13"
warn_return_any = false
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true
follow_imports = "skip"

[tool.ruff]
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I", "W", "UP"]
ignore = ["E501"]
